mfbas/requirements.txt
django==4.2.13
djangorestframework==3.15.1
facenet-pytorch==2.5.3
torch==2.0.1
torchvision==0.15.2
Pillow==10.3.0
bcrypt==4.1.3
cryptography==42.0.5
psycopg2-binary==2.9.9
python-dotenv==1.0.1
